+++ electron/src/main/settingsSnapshot.js
@@ -15,6 +15,8 @@ function readSettingsSnapshot(store, options = {}) {
     forwards: store.get('forwards'),
     flrigEnabled: store.get('flrigEnabled', store.get('rigctldEnabled', false)),
     flrigEndpoint: store.get('flrigEndpoint', store.get('rigctldEndpoint', '127.0.0.1:12345')),
+    defaultMyCall: store.get('defaultMyCall', ''),
+    defaultMyGrid: store.get('defaultMyGrid', ''),
     autoStartRelay: store.get('autoStartRelay', false),
     usePotaSpotMap: store.get('usePotaSpotMap', false),
     qrzLoggingEnabled: store.get('qrzLoggingEnabled', false),